+++ src/main/java/core/DBProvider.java
@@ -1,5 +1,6 @@
 package main.java.core;
 
+import de.maxikg.osuapi.model.Beatmap;
 import main.java.util.secrets;
 
 import java.sql.*;
@@ -52,14 +53,14 @@ public class DBProvider {
         return rs.getInt(mods);
     }
 
-    public static int getAverage(String mods) throws ClassNotFoundException, SQLException {
+    public static double getAverage(String mods) throws ClassNotFoundException, SQLException {
         mods = prepareMods(mods);
         Class.forName("com.mysql.cj.jdbc.Driver");
         Connection c = DriverManager.getConnection(secrets.dbPath, secrets.dbUser, secrets.dbPw);
         Statement stmnt = c.createStatement();
         ResultSet rs = stmnt.executeQuery("select avg(" + mods + ") as " + mods + " from ppRatings where " + mods + "!=-1");
         rs.next();
-        return rs.getInt(mods);
+        return rs.getDouble(mods);
     }
 
     public static void addMapPp(int mapID) throws ClassNotFoundException, SQLException {
